Capture errors and display it.

+++ deployment/deployInsightAgent.sh
@@ -46,18 +46,34 @@ if [ $AGENT_TYPE != 'proc' ] && [ $AGENT_TYPE != 'cadvisor' ] && [ $AGENT_TYPE !
 fi
 
 wget https://bootstrap.pypa.io/get-pip.py && python get-pip.py --force-reinstall --user
-wget --no-check-certificate https://raw.githubusercontent.com/insightfinder/InsightAgent/master/deployment/requirements
+wget --no-check-certificate https://raw.githubusercontent.com/insightfinder/InsightAgent/testing/deployment/requirements
 pip install -U --force-reinstall --user virtualenv
+if [ "$?" -ne "0" ]; then
+    echo "pip install failed. Please install the pre-requisites mentioned in README"
+    exit 1
+fi
 version=`python -c 'import sys; print(str(sys.version_info[0])+"."+str(sys.version_info[1]))'`
+if [ "$?" -ne "0" ]; then
+    echo "Unable to get python version. Please install the pre-requisites mentioned in README"
+    exit 1
+fi
 python  /home/$USER/.local/lib/python$version/site-packages/virtualenv.py pyenv
+if [ "$?" -ne "0" ]; then
+    echo "python virtual environment install failed. Please install the pre-requisites mentioned in README"
+    exit 1
+fi
 source pyenv/bin/activate
 pip install -r requirements
+if [ "$?" -ne "0" ]; then
+    echo "install failed. Please install the pre-requisites mentioned in README"
+    exit 1
+fi
 deactivate
 
 rm requirements
 rm get-pip.py
 
-wget --no-check-certificate https://raw.githubusercontent.com/insightfinder/InsightAgent/master/deployment/deployInsightAgent.py
+wget --no-check-certificate https://raw.githubusercontent.com/insightfinder/InsightAgent/testing/deployment/deployInsightAgent.py
 sudo python deployInsightAgent.py -n $INSIGHTFINDER_USERNAME -i $PROJECTNAME -u $USERNAME -k $LICENSEKEY -s $SAMPLING_INTERVAL -r $REPORTING_INTERVAL -t $AGENT_TYPE
 rm -rf pyenv
 rm deployInsightAgent.sh
